+++ src/components/Form.js
@@ -1,0 +1,11 @@
+const MyForm = () => (
+  <>
+    <input type="text" id="title" placeholder="Your Title Here..." />
+    <br />
+    <input type="text" id="name" placeholder="Your Name Here..." />
+    <br />
+    <button type="submit">Add Book</button>
+  </>
+);
+
+export default MyForm;
